Move AVR code out of Marlin Core
